squashfs-tools: Allow COMP_DEFAULT to be overridden via make cmdline
Currently the only way to modify the default compression algorithm is to
modify Makefile. It is useful to be able to set the default compression
algorithm via the make command line since otherwise a build time search
replace operation needs to be performed on the Makefile. My use case for
this is adding PACKAGECONFIG support to Yocto for selecting the default
compression algorithm.

Signed-off-by: Peter Morrow <pemorrow@linux.microsoft.com>